+++ src/boxes/font.js
@@ -1,0 +1,6 @@
+import { createTrackReferenceTypeBox } from "./helpers.js";
+
+export default createTrackReferenceTypeBox(
+  "Font Track Reference Type Box",
+  "Lists track IDs that carry font resources used by the containing track.",
+);