+++ src/boxes/index.js
@@ -4,6 +4,7 @@ import avc1 from "./avc1.js";
 import avc3 from "./avc3.js";
 import avcC from "./avcC.js";
 import btrt from "./btrt.js";
+import cdsc from "./cdsc.js";
 import co64 from "./co64.js";
 import colr from "./colr.js";
 import cslg from "./cslg.js";
@@ -21,11 +22,15 @@ import emsg from "./emsg.js";
 import enca from "./enca.js";
 import encv from "./encv.js";
 import esds from "./esds.js";
+import font from "./font.js";
 import free from "./free.js";
 import frma from "./frma.js";
 import ftyp from "./ftyp.js";
 import hdlr from "./hdlr.js";
 import hev1 from "./hev1.js";
+import hind from "./hind.js";
+import hint from "./hint.js";
+import hmhd from "./hmhd.js";
 import hvc1 from "./hvc1.js";
 import hvcC from "./hvcC.js";
 import ID32 from "./ID32.js";
@@ -40,6 +45,7 @@ import mehd from "./mehd.js";
 import meta from "./meta.js";
 import mfhd from "./mfhd.js";
 import mfra from "./mfra.js";
+import mfro from "./mfro.js";
 import minf from "./minf.js";
 import moof from "./moof.js";
 import moov from "./moov.js";
@@ -48,6 +54,7 @@ import mvex from "./mvex.js";
 import mvhd from "./mvhd.js";
 import nmhd from "./nmhd.js";
 import Opus from "./Opus.js";
+import padb from "./padb.js";
 import pasp from "./pasp.js";
 import pdin from "./pdin.js";
 import prft from "./prft.js";
@@ -66,19 +73,25 @@ import skip from "./skip.js";
 import smhd from "./smhd.js";
 import stbl from "./stbl.js";
 import stco from "./stco.js";
+import stdp from "./stdp.js";
 import sthd from "./sthd.js";
 import stsc from "./stsc.js";
 import stsd from "./stsd.js";
+import stsh from "./stsh.js";
 import stss from "./stss.js";
 import stsz from "./stsz.js";
 import stts from "./stts.js";
 import styp from "./styp.js";
+import stz2 from "./stz2.js";
+import subt from "./subt.js";
 import tenc from "./tenc.js";
 import tfdt from "./tfdt.js";
 import tfhd from "./tfhd.js";
+import tfra from "./tfra.js";
 import tkhd from "./tkhd.js";
 import traf from "./traf.js";
 import trak from "./trak.js";
+import tref from "./tref.js";
 import trep from "./trep.js";
 import trex from "./trex.js";
 import trun from "./trun.js";
@@ -86,7 +99,9 @@ import udta from "./udta.js";
 import url_ from "./url .js";
 import urn_ from "./urn .js";
 import uuid from "./uuid.js";
+import vdep from "./vdep.js";
 import vmhd from "./vmhd.js";
+import vplx from "./vplx.js";
 
 /**
  * @type {Record<string, import("./types.js").BoxDefinition<{ [key: string]: unknown }>>}
@@ -100,6 +115,7 @@ export default {
   avc3,
   avcC,
   btrt,
+  cdsc,
   colr,
   co64,
   cslg,
@@ -117,11 +133,15 @@ export default {
   enca,
   encv,
   esds,
+  font,
   free,
   frma,
   ftyp,
   hdlr,
   hev1,
+  hind,
+  hint,
+  hmhd,
   hvc1,
   hvcC,
   iods,
@@ -135,6 +155,7 @@ export default {
   meta,
   mfhd,
   mfra,
+  mfro,
   minf,
   moof,
   moov,
@@ -142,6 +163,7 @@ export default {
   mvex,
   mvhd,
   nmhd,
+  padb,
   pasp,
   pdin,
   pssh,
@@ -158,21 +180,27 @@ export default {
   sinf,
   smhd,
   skip,
+  stdp,
   sthd,
   stbl,
   stco,
   stsc,
   stsd,
+  stsh,
   stss,
   stsz,
   stts,
   styp,
+  stz2,
+  subt,
+  tfra,
   tfdt,
   tfhd,
   tenc,
   tkhd,
   traf,
   trak,
+  tref,
   trep,
   trex,
   trun,
@@ -180,5 +208,7 @@ export default {
   "url ": url_,
   "urn ": urn_,
   uuid,
+  vdep,
   vmhd,
+  vplx,
 };