+++ src/boxes/vdep.js
@@ -1,0 +1,6 @@
+import { createTrackReferenceTypeBox } from "./helpers.js";
+
+export default createTrackReferenceTypeBox(
+  "Auxiliary Depth Video Track Reference Type Box",
+  "Lists video tracks referenced by the containing auxiliary depth track.",
+);